update some field types and the references to them

+++ book-store/BookStore.Domain/Product.cs
@@ -4,13 +4,13 @@ namespace BookStore.Domain
 {
     public class Product
     {
-        public Product(string name, double price)
+        public Product(string name, decimal price)
         {
             Name = name;
             Price = price;
         }
 
-        public Product(int id, string name, double price)
+        public Product(int id, string name, decimal price)
         {
             ID = id;
             Name = name;
@@ -19,6 +19,6 @@ namespace BookStore.Domain
 
         public int ID {get; set;}
         public string Name {get; set;}
-        public double Price {get; set;}
+        public decimal Price {get; set;}
     }
 }
